backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.2.1
requests==2.31.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.12
